create new class directly in the module path
